requirements.txt
certifi==2017.7.27.1
Django==1.11.3
django-bootstrap-form==3.3
django-bootstrap-toolkit==2.15.0
django-bootstrap4==0.0.4
django-forms-bootstrap==3.1.0
django-registration-redux==1.8
olefile==0.44
Pillow==4.2.1
pytz==2017.2
wincertstore==0.2
